# Conflicts:
#	clients/pkg/logentry/stages/stage.go

+++ integration/cluster/cluster.go
@@ -1,14 +1,15 @@
 package cluster
 
 import (
+	"context"
 	"errors"
 	"flag"
 	"fmt"
 	"net"
 	"net/http"
 	"net/http/httptest"
-	"net/url"
 	"os"
+	"path/filepath"
 	"strings"
 	"sync"
 	"text/template"
@@ -16,9 +17,11 @@ import (
 
 	"github.com/grafana/dskit/multierror"
 	"github.com/prometheus/client_golang/prometheus"
+	"gopkg.in/yaml.v2"
 
 	"github.com/grafana/loki/pkg/loki"
 	"github.com/grafana/loki/pkg/util/cfg"
+	"github.com/grafana/loki/pkg/validation"
 )
 
 var (
@@ -28,8 +31,11 @@ var (
 auth_enabled: true
 
 server:
-  http_listen_port: {{.httpPort}}
-  grpc_listen_port: {{.grpcPort}}
+  http_listen_port: 0
+  grpc_listen_port: 0
+  grpc_server_max_recv_msg_size: 110485813
+  grpc_server_max_send_msg_size: 110485813
+
 
 common:
   path_prefix: {{.dataPath}}
@@ -42,6 +48,12 @@ common:
     instance_addr: 127.0.0.1
     kvstore:
       store: inmemory
+
+limits_config:
+  per_stream_rate_limit: 50MB
+  per_stream_rate_limit_burst: 50MB
+  ingestion_rate_mb: 50
+  ingestion_burst_size_mb: 50
 
 storage_config:
   boltdb_shipper:
@@ -71,11 +83,8 @@ ingester:
   lifecycler:
     min_ready_duration: 0s
 
-frontend_worker:
-  scheduler_address: localhost:{{.schedulerPort}}
-
-frontend:
-  scheduler_address: localhost:{{.schedulerPort}}
+querier:
+  multi_tenant_queries_enabled: true
 
 {{if .remoteWriteUrls}}
 ruler:
@@ -146,9 +155,10 @@ func (w *wrappedRegisterer) MustRegister(collectors ...prometheus.Collector) {
 }
 
 type Cluster struct {
-	sharedPath string
-	components []*Component
-	waitGroup  sync.WaitGroup
+	sharedPath    string
+	overridesFile string
+	components    []*Component
+	waitGroup     sync.WaitGroup
 }
 
 func New() *Cluster {
@@ -158,20 +168,48 @@ func New() *Cluster {
 		panic(err.Error())
 	}
 
+	overridesFile := filepath.Join(sharedPath, "loki-overrides.yaml")
+
+	err = os.WriteFile(filepath.Join(sharedPath, "loki-overrides.yaml"), []byte(`overrides:`), 0777)
+	if err != nil {
+		panic(fmt.Errorf("error creating overrides file: %w", err))
+	}
+
 	return &Cluster{
-		sharedPath: sharedPath,
+		sharedPath:    sharedPath,
+		overridesFile: overridesFile,
 	}
 }
 
 func (c *Cluster) Run() error {
 	for _, component := range c.components {
+		if component.running {
+			continue
+		}
+
 		if err := component.run(); err != nil {
 			return err
 		}
 	}
 	return nil
 }
+
+func (c *Cluster) Restart() error {
+	if err := c.stop(false); err != nil {
+		return err
+	}
+
+	return c.Run()
+}
+
 func (c *Cluster) Cleanup() error {
+	return c.stop(true)
+}
+
+func (c *Cluster) stop(cleanupFiles bool) error {
+	_, cancelFunc := context.WithTimeout(context.Background(), time.Second*3)
+	defer cancelFunc()
+
 	var (
 		files []string
 		dirs  []string
@@ -194,12 +232,14 @@ func (c *Cluster) Cleanup() error {
 	// wait for all process to close
 	c.waitGroup.Wait()
 
-	// cleanup dirs/files
-	for _, d := range dirs {
-		errs.Add(os.RemoveAll(d))
-	}
-	for _, f := range files {
-		errs.Add(os.Remove(f))
+	if cleanupFiles {
+		// cleanup dirs/files
+		for _, d := range dirs {
+			errs.Add(os.RemoveAll(d))
+		}
+		for _, f := range files {
+			errs.Add(os.Remove(f))
+		}
 	}
 
 	return errs.Err()
@@ -207,20 +247,11 @@ func (c *Cluster) Cleanup() error {
 
 func (c *Cluster) AddComponent(name string, flags ...string) *Component {
 	component := &Component{
-		name:    name,
-		cluster: c,
-		flags:   flags,
-	}
-
-	var err error
-	component.httpPort, err = getFreePort()
-	if err != nil {
-		panic(fmt.Errorf("error allocating HTTP port: %w", err))
-	}
-
-	component.grpcPort, err = getFreePort()
-	if err != nil {
-		panic(fmt.Errorf("error allocating GRPC port: %w", err))
+		name:          name,
+		cluster:       c,
+		flags:         flags,
+		running:       false,
+		overridesFile: c.overridesFile,
 	}
 
 	c.components = append(c.components, component)
@@ -233,30 +264,30 @@ type Component struct {
 	cluster *Cluster
 	flags   []string
 
-	httpPort int
-	grpcPort int
+	configFile    string
+	overridesFile string
+	dataPath      string
+	rulerWALPath  string
+	rulesPath     string
+	RulesTenant   string
 
-	configFile   string
-	dataPath     string
-	rulerWALPath string
-	rulesPath    string
-	RulesTenant  string
+	running bool
+	wg      sync.WaitGroup
 
 	RemoteWriteUrls []string
 }
 
-func (c *Component) HTTPURL() *url.URL {
-	return &url.URL{
-		Host:   fmt.Sprintf("localhost:%d", c.httpPort),
-		Scheme: "http",
-	}
+func (c *Component) HTTPURL() string {
+	return fmt.Sprintf("http://localhost:%s", port(c.loki.Server.HTTPListenAddr().String()))
 }
 
-func (c *Component) GRPCURL() *url.URL {
-	return &url.URL{
-		Host:   fmt.Sprintf("localhost:%d", c.grpcPort),
-		Scheme: "grpc",
-	}
+func (c *Component) GRPCURL() string {
+	return fmt.Sprintf("localhost:%s", port(c.loki.Server.GRPCListenAddr().String()))
+}
+
+func port(addr string) string {
+	parts := strings.Split(addr, ":")
+	return parts[len(parts)-1]
 }
 
 func (c *Component) writeConfig() error {
@@ -306,9 +337,6 @@ func (c *Component) writeConfig() error {
 	if err := configTemplate.Execute(configFile, map[string]interface{}{
 		"dataPath":        c.dataPath,
 		"sharedDataPath":  c.cluster.sharedPath,
-		"grpcPort":        c.grpcPort,
-		"httpPort":        c.httpPort,
-		"schedulerPort":   c.grpcPort,
 		"remoteWriteUrls": c.RemoteWriteUrls,
 		"rulesPath":       c.rulesPath,
 		"rulerWALPath":    c.rulerWALPath,
@@ -320,11 +348,12 @@ func (c *Component) writeConfig() error {
 		return fmt.Errorf("error closing config file: %w", err)
 	}
 	c.configFile = configFile.Name()
-
 	return nil
 }
 
 func (c *Component) run() error {
+	c.running = true
+
 	if err := c.writeConfig(); err != nil {
 		return err
 	}
@@ -337,6 +366,8 @@ func (c *Component) run() error {
 		c.flags,
 		"-config.file",
 		c.configFile,
+		"-limits.per-user-override-config",
+		c.overridesFile,
 	), flagset); err != nil {
 		return err
 	}
@@ -359,7 +390,7 @@ func (c *Component) run() error {
 	go func() {
 		for {
 			time.Sleep(time.Millisecond * 200)
-			if c.loki.Server.HTTP == nil {
+			if c.loki == nil || c.loki.Server == nil || c.loki.Server.HTTP == nil {
 				continue
 			}
 
@@ -375,8 +406,11 @@ func (c *Component) run() error {
 	}()
 
 	c.cluster.waitGroup.Add(1)
+	c.wg.Add(1)
+
 	go func() {
 		defer c.cluster.waitGroup.Done()
+		defer c.wg.Done()
 		err := c.loki.Run(loki.RunOpts{})
 		if err != nil {
 			newErr := fmt.Errorf("error starting component %v: %w", c.name, err)
@@ -396,20 +430,15 @@ func (c *Component) run() error {
 
 // cleanup calls the stop handler and returns files and directories to be cleaned up
 func (c *Component) cleanup() (files []string, dirs []string) {
-	if c.loki != nil {
+	if c.loki != nil && c.loki.SignalHandler != nil {
 		c.loki.SignalHandler.Stop()
+		c.running = false
 	}
 	if c.configFile != "" {
 		files = append(files, c.configFile)
 	}
 	if c.dataPath != "" {
 		dirs = append(dirs, c.dataPath)
-	}
-	if p := c.httpPort; p != 0 {
-		allocatedFreePorts.free(p)
-	}
-	if p := c.grpcPort; p != 0 {
-		allocatedFreePorts.free(p)
 	}
 	if c.rulerWALPath != "" {
 		dirs = append(dirs, c.rulerWALPath)
@@ -421,68 +450,48 @@ func (c *Component) cleanup() (files []string, dirs []string) {
 	return files, dirs
 }
 
-// keep track of previously allocated random ports, to ensure them not to clash
-var (
-	allocatedFreePorts = newAllocatedPorts()
-)
-
-type allocatedPorts struct {
-	m    map[int]struct{}
-	lock sync.Mutex
+func (c *Component) Restart() error {
+	c.cleanup()
+	c.wg.Wait()
+	return c.run()
 }
 
-func newAllocatedPorts() *allocatedPorts {
-	return &allocatedPorts{
-		m: make(map[int]struct{}),
+type runtimeConfigValues struct {
+	TenantLimits map[string]*validation.Limits `yaml:"overrides"`
+}
+
+func (c *Component) SetTenantLimits(tenant string, limits validation.Limits) error {
+	rcv := runtimeConfigValues{}
+	rcv.TenantLimits = c.loki.TenantLimits.AllByUserID()
+	if rcv.TenantLimits == nil {
+		rcv.TenantLimits = map[string]*validation.Limits{}
 	}
-}
+	rcv.TenantLimits[tenant] = &limits
 
-func (a *allocatedPorts) reserve(p int) (ok bool) {
-	a.lock.Lock()
-	defer a.lock.Unlock()
-	if _, exists := a.m[p]; exists {
-		return false
+	config, err := yaml.Marshal(rcv)
+	if err != nil {
+		return err
 	}
-	a.m[p] = struct{}{}
-	return true
+
+	return os.WriteFile(c.overridesFile, config, 0777)
 }
 
-func (a *allocatedPorts) free(p int) {
-	a.lock.Lock()
-	defer a.lock.Unlock()
-	delete(a.m, p)
-}
-
-func getFreePort() (port int, err error) {
-	var a *net.TCPAddr
-	if a, err = net.ResolveTCPAddr("tcp", "localhost:0"); err == nil {
-		var l *net.TCPListener
-		if l, err = net.ListenTCP("tcp", a); err == nil {
-			defer l.Close()
-			port := l.Addr().(*net.TCPAddr).Port
-
-			if !allocatedFreePorts.reserve(port) {
-				// port has been allocated before, try your luck again
-				return getFreePort()
-			}
-			return port, nil
-		}
+func (c *Component) GetTenantLimits(tenant string) validation.Limits {
+	limits := c.loki.TenantLimits.TenantLimits(tenant)
+	if limits == nil {
+		return c.loki.Cfg.LimitsConfig
 	}
-	return
+
+	return *limits
 }
 
 func NewRemoteWriteServer(handler *http.HandlerFunc) *httptest.Server {
-	server := httptest.NewUnstartedServer(*handler)
-	p, err := getFreePort()
-	if err != nil {
-		panic(fmt.Errorf("error allocating HTTP port: %w", err))
-	}
-
-	l, err := net.Listen("tcp", fmt.Sprintf("127.0.0.1:%d", p))
+	l, err := net.Listen("tcp", "127.0.0.1:0")
 	if err != nil {
 		panic(fmt.Errorf("failed to listen on: %v", err))
 	}
 
+	server := httptest.NewUnstartedServer(*handler)
 	server.Listener = l
 	server.Start()
